finnair-microservice/requirements-docker.txt
# Docker-specific requirements for Finnair microservice
# These packages are already installed in the main Dockerfile
# but listed here for reference and potential future use

# Core dependencies
undetected-chromedriver>=3.5.0
selenium>=4.15.0
python-dotenv>=1.0.0
supabase>=2.0.0

# Additional utilities
requests>=2.31.0
urllib3>=2.0.0

